except NotAuthorized when package can not be viewed
As a non authorized user (not logged in): when you are on the dataset-detail-view, you will get an error, when one of the associated showcases has an association with a deleted package that your (non-authenticated) user isn’t allowed to access.
this bugfix prevents that the showcases, that can not be displayed due to the lack of Authorization, to be added to the view.