add compatibility for FontAwesome 4.x
The function _create_link_text (called in link_for) adds the fontawesome-prefix ‘fa fa-‘ to the value that is set in the icon-attribute by just prepending it to the html-class-attribute.
That is why, for compatibility for FontAwesome 3.x, the former icon-names are added to the icon-attribute as a whole (with the former FontAwesome-Prefix ‘icon-’) at the end of the icon-string.