Use fq for dataset search, not q
The Solr maxBooleanClauses config only applies to terms in a query string, not a filter query: https://solr.apache.org/guide/8_1/query-settings-in-solrconfig.html#maxbooleanclauses

This will prevent Solr crashing when a showcase has >1024 datasets associated with it.

When searching for datasets that are *not* associated wtih a showcase, we already use an equivalent query in the fq parameter: https://github.com/ckan/ckanext-showcase/blob/master/ckanext/showcase/utils.py#L313.